+++ src/main/java/com/offbytwo/jenkins/model/AllBuilds.java
@@ -1,0 +1,28 @@
+/*
+ * Copyright (c) 2013 Rising Oak LLC.
+ *
+ * Distributed under the MIT license: http://opensource.org/licenses/MIT
+ */
+
+package com.offbytwo.jenkins.model;
+
+import java.util.List;
+
+/**
+ * This class is only needed to get all builds in
+ * {@link JobWithDetails#getAllBuilds()}.
+ * 
+ * @author Karl Heinz Marbaise
+ *
+ *         NOTE: This class is not part of any public API
+ */
+class AllBuilds extends BaseModel {
+    private List<Build> allBuilds;
+
+    public AllBuilds() {
+    }
+
+    public List<Build> getAllBuilds() {
+        return this.allBuilds;
+    }
+}